moving ARnftcore insid code folder - doesn't change anything, issue remain
